+++ lib/src/bot.dart
@@ -13,7 +13,7 @@ import 'package:weather/src/modules/weather/weather_manager.dart';
 import 'package:weather/src/modules/panorama/panorama_manager.dart';
 import 'package:weather/src/modules/dadjokes/dadjokes_manager.dart';
 import 'package:weather/src/modules/reputation/reputation_manager.dart';
-import 'package:weather/src/modules/youtube_manager.dart';
+import 'package:weather/src/modules/youtube/youtube_manager.dart';
 import 'package:weather/src/modules/conversator/conversator_manager.dart';
 import 'package:weather/src/modules/general/general_manager.dart';
 import 'package:weather/src/modules/accordion_poll/accordion_poll_manager.dart';
